Allow to skip {course_dir} in the exchange folder: place {inbound,outboud,feedback} directly in course_dir
When working with k8s, or systems where we mount folders indivudally,
we don't need to have a /exchange/{course_id}/inbound/{...} structure,
but {inbound,outbound,assignments} can live directly in the course root.

This simplifies permissions setup in some setups: we don't want to rely
on the container runtime to create the tree
/nbgrader/{course}/inbound/{user} for us, since the
/nbgrader/{course}/inbound may end up with wrong permissions (e.g. it
could get created as root with permission 700, and we would later need
some trick to change its permissions).